+++ src/core.c
@@ -634,3 +634,20 @@ inline void nano_ReleaseObject(SEXP x) {
   SETCAR(CDR(x), CAR(x));
 
 }
+
+void raio_invoke_cb(void *arg) {
+
+  SEXP call, context, data, node = (SEXP) arg, ctx = TAG(node);
+  context = Rf_findVarInFrame(ctx, nano_ContextSymbol);
+  if (context == R_UnboundValue) {
+    SET_TAG(node, R_NilValue);
+    return;
+  }
+  PROTECT(context);
+  data = rnng_aio_get_msg(context);
+  PROTECT(call = Rf_lcons(nano_ResolveSymbol, Rf_cons(data, R_NilValue)));
+  Rf_eval(call, ctx);
+  UNPROTECT(2);
+  // unreliable to release linked list node from later cb, just free the payload
+  SET_TAG(node, R_NilValue);
+}